Allow JWTs to contain non-string fields.
Trying to unpack a JWT containing an object field ("jwk") seems to fail
right now. This is because this code doesn't use an explicit struct to
unpack the JSON, but wants to unpack it into a map.

Change this code to be a bit more lenient by unpacking into a map of
empty interfaces. Convert values to strings when needed.